Move add_handler to gramex.transforms.handler
Also allow POSTing queryparams and path args
